[REFACTOR] Services are moved to your directory and all the components that import it are modified
