+++ src/app/header/header.component.html
@@ -4,7 +4,7 @@
             <i class="material-icons toCenter">menu</i>
         </div>
         <h1 class="logo opt-box" (click)="navigateTo('home')">Filmist</h1>
-        <div class="version">V12</div>
+        <!-- <div class="version">V12</div> -->
         <!--<div class="options login opt-box" (click)="showMenuOptions()" *ngIf="!showMiniLogin">-->
         <div class="options login opt-box" routerLink="/login" *ngIf="!showMiniLogin">
             <i class="material-icons">vpn_key</i>
