+++ src/app/video-genre/video-genre.resolve.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 
-import { DbmoviesService } from '../dbmovies.service';
+import { DbmoviesService } from '../../services/dbmovies.service';
 
 @Injectable()
 export class MovieGenreResolve implements Resolve<any> {
